tim_service/requirements.txt
# TIM Motion Service Requirements
# =================================
# Install with: pip install -r requirements.txt

# Core dependencies
pyyaml>=6.0              # Configuration file parsing
numpy>=1.22.0            # RapidCode requirement (if using real hardware)

# Optional: for real RapidCode (installed separately from RSI)
# RSI RapidCode should be installed at C:\RSI\10.7.1 (Windows) or equivalent

# Testing & development
pytest>=7.0              # Unit testing
pytest-cov>=3.0          # Coverage reporting

# Logging & monitoring
python-daemon>=2.3       # For Windows Service integration
pywin32>=300             # Windows-specific tools (for service installation)
